Small issue with speaks not updating
Speaks has been updated from real data type to float data type to six small bug of speaks not updating through the Results page.

+++ PersonalJob_Reg.php
@@ -34,12 +34,12 @@ if (isset($_POST["btnBeginT"])) {
   	for ($Counter=1; $Counter <= $NumRounds; $Counter++) { 
   				$RoundName = "Round" . strval($Counter);
   				$sql = "ALTER TABLE speaks
-						ADD $RoundName numeric(3,2)";
+						ADD $RoundName FLOAT";
 				$result = $conn->query($sql);		
 
   			}
   	$sql = "ALTER TABLE speaks
-  			ADD Average numeric(3,2)";
+  			ADD Average FLOAT";
   	$result = $conn->query($sql);
 
   	$sql = "SELECT COUNT(DISTINCT TeamName) AS num FROM $TName";
@@ -65,7 +65,7 @@ if (isset($_POST["btnBeginT"])) {
  			ADD Wins tinyint";
  	$result = $conn->query($sql);
  	$sql = "ALTER TABLE wins
- 			ADD TotalScore numeric(3,2)";
+ 			ADD TotalScore FLOAT";
  	$result = $conn->query($sql);						
   	header("Location:PersonalJob_SeedRound.php");
